Clicking Creates Circles
Now two or three things happen.
First, it doesn't automaticall quit after 30 seconds anymore.
Second, it quits nicely when you ask it to.
Third, it creates a circle where you click.

This has demonstrated an issue with how I do collision detection.
I forgot to consider that circles can be within circles...
This could cause issues both with the way I'm currently doing it and the way I was planning on doing it.

So, that'll just stay an issue until I figure something else out.

Also, I'm drawing the newest circles first.
This means that when we create a circle within a circle it sits behind the other one.